Added support for ES module system for nanoid package to work
I added a way to support ES module system for nanoid package to be used. Since the code also consists smpp package which uses CommonJS way of module system, adding createRequire(import.meta.url); is important to make it work. 

In addition to that, modern js syntax replaced the old way of declaring and assigning variables.